Toezichter
Data, radiobuttons en combobox.

+++ SprintPlannerZM.Model/Deadline.cs
@@ -4,7 +4,7 @@ namespace SprintPlannerZM.Model
 {
     public class Deadline
     {
-        public int deadlineID { get; set; }
+        public byte deadlineID { get; set; }
         public DateTime deadlineLeerling { get; set; }
         public DateTime deadlineExamen { get; set; }
     }
